1、完成了智能训练模块的智能训练生成的分页查询 2、完成了创建随机答卷的接口 3、往答卷表添加了试卷名称的字段 4、完成查询随机试卷详情的接口 5、在设计提交答卷的接口时，发现可以调用原先的方法，故实现了封装提交答卷 6、在答卷表新增字段提交时间

+++ ms_stu_pro_320_server_new/ruoyi-manage/src/main/java/com/mashang/domain/entity/QuestionAnswer.java
@@ -35,7 +35,7 @@ public class QuestionAnswer extends BaseModel {
     private Long userId;
 
     @ApiModelProperty(value = "答卷id")
-    private Integer testAnswerId;
+    private Long testAnswerId;
 
     @ApiModelProperty(value = "题目题干")
     private String questionTitle;
